+++ apps/native-component-list/src/screens/UI/UIScreen.android.tsx
@@ -211,6 +211,14 @@ export const UIScreens = [
     },
   },
   {
+    name: 'RangeSlider component',
+    route: 'ui/range-slider',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./RangeSliderScreen'));
+    },
+  },
+  {
     name: 'DropdownMenu component',
     route: 'ui/dropdown-menu',
     options: {},