+++ packages/expo-ui/android/src/main/java/expo/modules/ui/ExpoUIModule.kt
@@ -316,6 +316,14 @@ class ExpoUIModule : Module() {
       }
     }
 
+    ExpoUIView<RangeSliderProps>("RangeSliderView") {
+      Events("onValueChange", "onValueChangeFinished")
+
+      Content { props ->
+        RangeSliderContent(props)
+      }
+    }
+
     ExpoUIView<ShapeProps>("ShapeView") {
       Content { props ->
         ShapeContent(props)